requirements.txt
anyio>=3.6.1
Nintendo>=0.0.0
anynet>=0.0.14
discord.py>=1.7.3
matplotlib>=3.5.2
pandas>=1.4.2
python-dotenv>=0.20.0